refactor(system): nest watcher API under Path
