+++ kyo-system/shared/src/main/scala/kyo/PathWatch.scala
@@ -1,5 +1,9 @@
 package kyo
 
+import kyo.Path.Change as PathChange
+import kyo.Path.MatchCase
+import kyo.Path.WatchDepth
+import kyo.Path.WatchOptions
 import kyo.kernel.ArrowEffect
 
 /** Capability required to acquire filesystem watchers.
@@ -13,75 +17,6 @@ import kyo.kernel.ArrowEffect
   * backends can expose observation independently of reading and writing.
   */
 sealed trait PathWatch extends ArrowEffect[[A] =>> Path.WatchOp[A], Id]
-
-/** Selects how deeply a watcher observes a directory.
-  *
-  * [[Immediate]] observes only direct children of the watched root.
-  * [[Recursive]] also observes descendants at every depth.
-  * The watched root itself is not passed through the glob filter.
-  * Its disappearance is instead reported as [[PathChange.Invalidated]].
-  *
-  * Depth is applied before glob matching, and paths emitted by either
-  * mode retain their complete backend path.
-  */
-enum WatchDepth derives CanEqual:
-    case Immediate
-    case Recursive
-end WatchDepth
-
-/** Selects the case policy used when matching a watch glob.
-  *
-  * [[FileSystemDefault]] delegates to the backend's native policy.
-  * [[Sensitive]] compares every glob component case sensitively.
-  * [[Insensitive]] compares every glob component without case.
-  *
-  * This setting changes only event selection. It does not rewrite
-  * the paths contained in emitted [[PathChange]] values.
-  */
-enum MatchCase derives CanEqual:
-    case FileSystemDefault
-    case Sensitive
-    case Insensitive
-end MatchCase
-
-/** A normalized filesystem change emitted by a [[Path.Watcher]].
-  *
-  * Creation, modification, and removal events identify one path.
-  * A move identifies both its former and current paths. Moves crossing
-  * a watch filter boundary are normalized to removal or creation.
-  * [[Overflow]] reports that bounded event delivery lost detail, while
-  * [[Invalidated]] reports that the watched root is no longer usable.
-  *
-  * Paths use the namespace of the filesystem that acquired the watcher.
-  */
-enum PathChange derives CanEqual:
-    case Created(path: Path)
-    case Modified(path: Path)
-    case Removed(path: Path)
-    case Moved(from: Path, to: Path)
-    case Overflow(root: Path)
-    case Invalidated(root: Path)
-end PathChange
-
-/** Configures filesystem watching and event selection.
-  *
-  * `depth` controls traversal below the watched root. `glob` is matched
-  * against paths relative to that root using `caseSensitivity`.
-  * `capacity` bounds pending changes and must be positive. When it is
-  * exceeded, queued detail is replaced by [[PathChange.Overflow]].
-  * `followLinks` controls whether linked directories and their targets
-  * participate in observation.
-  *
-  * The defaults observe direct children, match all names using the
-  * backend's case policy, retain 256 pending changes, and do not follow links.
-  */
-final case class WatchOptions(
-    depth: WatchDepth = WatchDepth.Immediate,
-    glob: Glob = Glob.all,
-    caseSensitivity: MatchCase = MatchCase.FileSystemDefault,
-    capacity: Int = 256,
-    followLinks: Boolean = false
-) derives CanEqual
 
 private[kyo] object PathWatch:
